# frontend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations": {
      "post": {
        "operationId": "OrganizationsController_create",
        "summary": "Create a new organization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Admin"]
      },
      "get": {
        "operationId": "OrganizationsController_getPaginated",
        "summary": "Get all organizations",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter criteria",
            "schema": {
              "type": "object"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": ["ASC", "DESC"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved all organizations successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Admin"]
      }
    },
    "/api/v1/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findOne",
        "summary": "Get an organization by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved organization successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Organization"]
      },
      "put": {
        "operationId": "OrganizationsController_update",
        "summary": "Update an organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Admin"]
      },
      "patch": {
        "operationId": "OrganizationsController_partialUpdate",
        "summary": "Partially update an organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization partially updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Admin"]
      },
      "delete": {
        "operationId": "OrganizationsController_remove",
        "summary": "Soft delete an organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Admin"]
      }
    },
    "/api/v1/organizations/{id}/restore": {
      "post": {
        "operationId": "OrganizationsController_restore",
        "summary": "Restore a soft-deleted organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Organization restored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Admin"]
      }
    },
    "/api/v1/organizations/{id}/users/{userId}": {
      "post": {
        "operationId": "OrganizationsController_addUserToOrganization",
        "summary": "Add a user to an organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Added user to organization successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SafeUserDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Organization"]
      },
      "delete": {
        "operationId": "OrganizationsController_removeUserFromOrganization",
        "summary": "Remove user from an organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed user from organization successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SafeUserDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Organization"]
      }
    },
    "/api/v1/organizations/{id}/users": {
      "get": {
        "operationId": "OrganizationsController_getPaginatedUsers",
        "summary": "Get all users",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter criteria",
            "schema": {
              "type": "object"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": ["ASC", "DESC"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved organization users successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Organization"]
      }
    },
    "/api/v1/organizations/{id}/groups": {
      "get": {
        "operationId": "OrganizationsController_getOrganizationGroups",
        "summary": "Get all groups of an organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved organization groups successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SafeGroupDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Organization"]
      }
    },
    "/api/v1/organizations/{id}/employees": {
      "get": {
        "operationId": "OrganizationsController_getOrganizationEmployees",
        "summary": "Get all employees of an organization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved organization employees successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SafeEmployeeDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Content not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": ["organizations", "Organization"]
      }
    },
...
  },
  "info": {
    "title": "OpenAPI",
    "description": "\n      # Authentication and Permissions\n\n      This API uses the following authentication and permission levels:\n\n      - **Public**: No authentication required\n      - **User** \n      - **Admin**\n      - **Organization Member**\n\n      Each endpoint in this documentation specifies its required permission level in the description.\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "user_sid"
      }
    },
    "schemas": {
...
      "SafeUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          }
        },
        "required": ["id", "email", "firstName", "lastName", "organizationId"]
      },
...
      "PaginationResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "pages": {
            "type": "number"
          }
        },
        "required": ["items", "total", "page", "limit", "pages"]
      },
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["code", "name"]
      },
      "SafeOrganizationDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["code", "name"]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["code", "name"]
      },
      "SafeGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          }
        },
        "required": ["id", "name", "organizationId"]
      },
      "SafeEmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          }
        },
        "required": ["id", "name", "phone_number", "organizationId"]
      },
...
    }
  }
}